# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: KwikFinder API
  version: "1.0.0"
  description: |
    Read-only store catalog API. Clients poll `/v1/meta` and download the full
    catalog from `/v1/stores` when `dataVersion` changes. No write endpoints
    and no API keys for v1.
servers:
  - url: http://localhost:8080
    description: Local / Docker Compose default
paths:
...
  /v1/stores/{id}:
    get:
      tags: [stores]
      summary: Single store by id
      operationId: getStore
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Store id (Kwik Trip store number)
      responses:
        "200":
          description: Store object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Store:
      type: object
      required:
        - id
        - name
        - latitude
        - longitude
        - address1
        - city
        - county
        - state
        - zip
        - phone
        - open24Hours
        - hours
        - fuels
        - amenities
        - truckParkingSpaces
        - familyRestroom
        - evCharging
        - features
      properties:
        id:
          type: integer
        name:
          type: string
          example: "KWIK TRIP #123"
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        address1:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        county:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        open24Hours:
          type: boolean
        hours:
          type: array
          items:
            $ref: "#/components/schemas/StoreHours"
        fuels:
          type: array
          items:
            $ref: "#/components/schemas/FuelOffering"
        amenities:
          type: array
          items:
            type: string
        truckParkingSpaces:
          type: integer
          minimum: 0
        familyRestroom:
          type: boolean
        evCharging:
          type: string
          nullable: true
          enum: [open, comingSoon]
          description: null | open | comingSoon (from Maps & Downloads EV PDF)
        features:
          $ref: "#/components/schemas/StoreFeatures"
    StoreHours:
      type: object
      required:
        - dayOfWeek
        - openTime
        - closeTime
      properties:
        dayOfWeek:
          type: string
          example: Monday
        openTime:
          type: string
          example: "00:00:00"
        closeTime:
          type: string
          example: "00:00:00"
    FuelOffering:
      type: object
      required:
        - type
        - description
        - price
      properties:
        type:
          type: string
          example: "DIESEL #2"
        description:
          type: string
          nullable: true
        price:
          type: number
          format: double
          nullable: true
    StoreFeatures:
      type: object
      description: |
        Server-owned filter flags derived from fuels, amenities, and related
        fields. Clients filter on these instead of re-mapping Kwik Trip strings.
      properties:
        diesel:
          type: boolean
        premiumDiesel:
          type: boolean
        def:
          type: boolean
        e85:
          type: boolean
        cng:
          type: boolean
        noEthanolGas:
          type: boolean
        unleaded88:
          type: boolean
        scale:
          type: boolean
        showers:
          type: boolean
        truckParking:
          type: boolean
        truckParkingStalls:
          type: integer
          minimum: 0
        transFlo:
          type: boolean
        fleetCards:
          type: boolean
        truckFriendly:
          type: boolean
        carWash:
          type: boolean
        atm:
          type: boolean
        bitcoinATM:
          type: boolean
        wifi:
          type: boolean
        restaurant:
          type: boolean
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: Store not found